Worked on scripting and Level 1
Jhoxua I organized some of the stuff in your level. I also added the two enemies in the beginning so you can put them throughout the map. Make sure that you fix the gaps in the maze. There are many imperfections in it so try to make it look as good as possible.